develop の main.rs への wind 系 CLI override 追加を bootstrap.rs の
apply_engine_overrides へ移植し、main.rs は thin main を維持して解決。

+++ blender_addon/effects/flame/tests/test_properties.py
@@ -56,7 +56,7 @@ class TestPropertyKind:
         assert property_kind(1.5) == "float"
 
     def test_int_value(self):
-        assert property_kind(42) == "float"
+        assert property_kind(42) == "int"
 
 
 class TestCollectParams:
